% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{economy}
\alias{economy}
\alias{economy_text}
\alias{economy_dates}
\alias{text}
\alias{dates}
\title{A Snippet of the Economy Dataset from toscaData}
\format{
\code{economy_text} is a list of tokenized texts of length 576.

\code{economy_dates} is

An object of class \code{Date} of length 576.
}
\source{
\url{https://github.com/Docma-TU/toscaData}
}
\usage{
data(economy_text)

data(economy_dates)
}
\description{
Example Dataset from Wikinews consisting of 576 articles. It can be used to
familiarize with the functions offered by this package.
}
\keyword{datasets}
