% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.RollingLDA.R
\name{as.RollingLDA}
\alias{as.RollingLDA}
\alias{is.RollingLDA}
\title{RollingLDA Object}
\usage{
as.RollingLDA(x)

is.RollingLDA(obj, verbose = FALSE)
}
\arguments{
\item{x}{[\code{named list}]\cr
Output from \code{\link[lda]{lda.collapsed.gibbs.sampler}}. Alternatively each
element can be passed for individual results. Individually set elements
overwrite elements from \code{x}.}

\item{obj}{[\code{R} object]\cr
Object to test.}

\item{verbose}{[\code{logical(1)}]\cr
Should test information be given in the console?}

\item{id}{TBA}

\item{lda}{TBA}

\item{docs}{TBA}

\item{dates}{TBA}

\item{vocab}{TBA}

\item{chunks}{TBA}

\item{param}{[\code{named list}]\cr
Parameters of the function call \code{\link[lda]{lda.collapsed.gibbs.sampler}}.
List always should contain names "K", "alpha", "eta" and "num.iterations".}
}
\value{
[\code{named list}] \code{\link{RollingLDA}} object.
}
\description{
Constructor for RollingLDA objects used in this package.
}
\details{
If you call \code{as.RollingLDA} on an object \code{x} which already is of
the structure of an \code{RollingLDA} object (in particular a \code{RollingLDA}
object itself), the additional arguments \code{id, param, ...}
may be used to override the specific elements.
}
\seealso{
Other RollingLDA functions: 
\code{\link{RollingLDA}()},
\code{\link{getChunks}()},
\code{\link{updateRollingLDA}()}
}
\concept{RollingLDA functions}
